refactor: extract KadDHT specific analysis logic from Nimlibp2pAnalyzer into a dedicated KadDHTAnalyzer class
